24/2/17
도감인덱스 스크립트 삭제 도감씬 없애 예정 도감 스테이지 캔버스에 적용
