--- Assets/Cjy/Script/Book/IndexCanvusButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2d4c438846ee2984e9d0692c501612c9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
